test/com/twu/biblioteca/BookTest.java
package com.twu.biblioteca;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;

import java.io.ByteArrayOutputStream;
import java.io.PrintStream;

import static org.junit.Assert.assertEquals;

public class BookTest {

    private final ByteArrayOutputStream outContent = new ByteArrayOutputStream();
    @Before
    public void setUpStreams() {
        System.setOut(new PrintStream(outContent));
    }

    @Test
    public void specDisplayingABook() {
        Book book = new Book("Alchemist", "Paulo Coelho", 2011);
        book.display();
        assertEquals(String.format("%-23s%-23s%-23d\n", "Alchemist", "Paulo Coelho", 2011), outContent.toString());
    }

    @After
    public void cleanUpStreams() {
        System.setOut(System.out);
    }
}